Move shared batch functionality
