+++ src/vercel/_internal/unstable/sandbox/runtime_common.py
@@ -3,7 +3,7 @@
 import copy
 import posixpath
 import signal as signal_module
-from collections.abc import Callable
+from collections.abc import Callable, Sequence
 from dataclasses import replace
 from datetime import timedelta
 from enum import Enum, auto
@@ -16,6 +16,7 @@ from vercel._internal.unstable.sandbox.models import (
     JSONValue,
     ProcessStatus,
     SandboxStatus,
+    _WriteFile,
 )
 from vercel._internal.unstable.sandbox.state import (
     ProcessState,
@@ -34,6 +35,48 @@ class _FilesystemBatchState(Enum):
     CREATED = auto()
     ACTIVE = auto()
     CLOSED = auto()
+
+
+class _SandboxFilesystemBatchBase:
+    __slots__ = ("_files", "_state")
+
+    def __init__(self) -> None:
+        self._files: list[_WriteFile] = []
+        self._state = _FilesystemBatchState.CREATED
+
+    def _stage(self, file: _WriteFile) -> None:
+        if self._state is not _FilesystemBatchState.ACTIVE:
+            raise RuntimeError("filesystem batch staging is only allowed inside its context")
+        self._files.append(file)
+
+    def write_bytes(self, path: RemotePath, data: bytes, *, mode: int | None = None) -> None:
+        self._stage(_WriteFile(path=_coerce_remote_path(path), content=data, mode=mode))
+
+    def write_text(
+        self,
+        path: RemotePath,
+        text: str,
+        *,
+        encoding: str = "utf-8",
+        errors: str = "strict",
+        mode: int | None = None,
+    ) -> None:
+        self._stage(
+            _WriteFile(
+                path=_coerce_remote_path(path),
+                content=text.encode(encoding, errors=errors),
+                mode=mode,
+            )
+        )
+
+    def _enter(self) -> None:
+        if self._state is not _FilesystemBatchState.CREATED:
+            raise RuntimeError("filesystem batch contexts can only be entered once")
+        self._state = _FilesystemBatchState.ACTIVE
+
+    def _close(self) -> Sequence[_WriteFile]:
+        self._state = _FilesystemBatchState.CLOSED
+        return self._files
 
 
 def _coerce_remote_path(path: RemotePath) -> str:
