+++ src/vercel/_internal/unstable/sandbox/sync_runtime.py
@@ -46,8 +46,8 @@ from vercel._internal.unstable.sandbox.runtime_common import (
     SandboxHandleBase,
     SnapshotHandleBase,
     _coerce_remote_path,
-    _FilesystemBatchState,
     _ProcessHandleState,
+    _SandboxFilesystemBatchBase,
     _signal_number,
 )
 from vercel._internal.unstable.sandbox.service import SandboxService, _SandboxTerminalState
@@ -319,43 +319,15 @@ class SyncSandboxFilesystem:
         )
 
 
-class SyncSandboxFilesystemBatch:
-    __slots__ = ("_files", "_state", "_write_files")
+class SyncSandboxFilesystemBatch(_SandboxFilesystemBatchBase):
+    __slots__ = ("_write_files",)
 
     def __init__(self, *, write_files: Callable[[Sequence[_WriteFile]], None]) -> None:
+        super().__init__()
         self._write_files = write_files
-        self._files: list[_WriteFile] = []
-        self._state = _FilesystemBatchState.CREATED
-
-    def _stage(self, file: _WriteFile) -> None:
-        if self._state is not _FilesystemBatchState.ACTIVE:
-            raise RuntimeError("filesystem batch staging is only allowed inside its context")
-        self._files.append(file)
-
-    def write_bytes(self, path: RemotePath, data: bytes, *, mode: int | None = None) -> None:
-        self._stage(_WriteFile(path=_coerce_remote_path(path), content=data, mode=mode))
-
-    def write_text(
-        self,
-        path: RemotePath,
-        text: str,
-        *,
-        encoding: str = "utf-8",
-        errors: str = "strict",
-        mode: int | None = None,
-    ) -> None:
-        self._stage(
-            _WriteFile(
-                path=_coerce_remote_path(path),
-                content=text.encode(encoding, errors=errors),
-                mode=mode,
-            )
-        )
 
     def __enter__(self) -> "SyncSandboxFilesystemBatch":
-        if self._state is not _FilesystemBatchState.CREATED:
-            raise RuntimeError("filesystem batch contexts can only be entered once")
-        self._state = _FilesystemBatchState.ACTIVE
+        self._enter()
         return self
 
     def __exit__(
@@ -364,9 +336,9 @@ class SyncSandboxFilesystemBatch:
         exc: BaseException | None,
         traceback: TracebackType | None,
     ) -> None:
-        self._state = _FilesystemBatchState.CLOSED
-        if exc_type is None and self._files:
-            self._write_files(self._files)
+        files = self._close()
+        if exc_type is None and files:
+            self._write_files(files)
 
 
 class SyncSandboxRuntimeSession(RuntimeSessionHandleBase):
